Add ability to specify allocation size limits
Sometimes it's not important how many objects are allocated, but how
much memory in bytes is being allocated. For now it's only possible to
limit the total memory allocated for all objects, even though we're
internally tracking it per object type.

This matcher is useful when dealing with large streams of data, for
example accepting large file uploads, and you want to assert that you're
keeping memory usage low with diligent string object management.